优化系统进程控制API

+++ django_common_task_system/views.py
@@ -10,6 +10,7 @@ from django_common_objects.models import CommonCategory
 from jionlp_time import parse_time
 from rest_framework import status
 from rest_framework.decorators import api_view
+from rest_framework.exceptions import NotFound
 from rest_framework.generics import CreateAPIView, RetrieveAPIView
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -101,7 +102,6 @@ def add_client(sender, instance: models.TaskClient, created, **kwargs):
         development server in single-threaded mode:
         $ python manage.py  runserver --nothreading --noreload
     """
-
 
 # for sig in [signal.SIGTERM, signal.SIGINT, getattr(signal, 'SIGQUIT', None), getattr(signal, 'SIGHUP', None)]:
 #     if sig is not None:
@@ -317,52 +317,61 @@ class ScheduleClientView:
     @api_view(['GET'])
     def action(request: Request, action: str):
         if action == 'start':
-            return ScheduleClientView.start_client()
+            client = TaskClient.objects.create()
+            return Response({'client_id': client.id})
+        if action == 'register':
+            return ScheduleClientView.register_client(request)
         client_id = request.GET.get('client_id', '')
         if not client_id.isdigit():
             return Response({'error': 'invalid client_id: %s' % client_id}, status=status.HTTP_400_BAD_REQUEST)
+        client: TaskClient = TaskClient.objects.get(int(client_id))
+        if client is None:
+            raise NotFound('TaskClient(%s)不存在' % client_id)
         if action == 'stop':
-            return ScheduleClientView.stop_client(int(client_id))
+            client.delete()
+            return Response({'message': 'stop client(%s) success' % client_id})
         elif action == 'log':
-            return ScheduleClientView.show_logs(int(client_id))
+            return HttpResponse(client.runner.read_log(), content_type='text/plain; charset=utf-8')
         else:
             return Response({'error': 'invalid action: %s, only support start/stop/log' % action},
                             status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
-    def show_logs(client_id: int):
-        # 此处pk为进程id
-        client: TaskClient = TaskClient.objects.get(client_id)
-        if client is None:
-            return HttpResponse('TaskClient(%s)不存在' % client_id)
-        if client.startup_status != TaskClientStatus.SUCCEED:
-            return HttpResponse(client.startup_log, content_type='text/plain; charset=utf-8')
-        logs = client.container.logs(tail=1000)
-        return HttpResponse(logs, content_type='text/plain; charset=utf-8')
-
-    @staticmethod
-    def report_client():
-        pass
-
-    @staticmethod
-    def start_client():
-        try:
-            client = TaskClient.objects.create()
-            return Response('TaskClient(%s)启动成功' % client.id)
-        except Exception as e:
-            return Response('启动TaskClient失败: %s' % e, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    @staticmethod
-    def stop_client(client_id: int):
-        client = TaskClient.objects.get(pk=client_id)
-        if client is None:
-            return Response('TaskClient(%s)不存在' % client_id, status=status.HTTP_404_NOT_FOUND)
-        try:
-            client.delete()
-            return Response('TaskClient(%s)已停止' % client_id)
-        except Exception as e:
-            return Response('停止TaskClient(%s)失败: %s' % (client_id, e),
-                            status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    def register_client(request: Request):
+        """
+        container: Container = None
+        group = models.CharField(max_length=100, verbose_name='分组')
+        subscription_url = models.CharField(max_length=200, verbose_name='订阅地址')
+        subscription_kwargs = models.JSONField(verbose_name='订阅参数', default=dict)
+        client_id = models.IntegerField(verbose_name='客户端ID', primary_key=True, default=0)
+        process_id = models.PositiveIntegerField(verbose_name='进程ID', null=True, blank=True)
+        container_id = models.CharField(max_length=100, verbose_name='容器ID', blank=True, null=True)
+        container_name = models.CharField(max_length=100, verbose_name='容器名称', blank=True, null=True)
+        container_image = models.CharField(max_length=100, verbose_name='容器镜像', blank=True, null=True)
+        container_status = models.CharField(choices=ContainerStatus.choices, default=ContainerStatus.NONE,
+                                            max_length=20, verbose_name='容器状态')
+        run_in_container = models.BooleanField(default=True, verbose_name='是否在容器中运行')
+        env = models.CharField(max_length=500, verbose_name='环境变量', blank=True, null=True)
+        startup_status = models.CharField(max_length=500, choices=TaskClientStatus.choices,
+                                          verbose_name='启动结果', default=TaskClientStatus.SUCCEED)
+        settings = models.TextField(verbose_name='配置', blank=True, null=True)
+        create_time = models.DateTimeField(auto_now_add=True, verbose_name='创建时间')
+        startup_log = models.CharField(max_length=2000, null=True, blank=True)
+        """
+        data = request.data
+        settings = data.get('settings')
+        container_id = data.get('container_id')
+        subscription_url = data.get('subscription_url')
+        subscription_kwargs = data.get('subscription_kwargs')
+        process_id = data.get('process_id')
+        client = TaskClient(container_id=container_id,
+                            subscription_url=subscription_url,
+                            startup_log='启动成功',
+                            settings=settings,
+                            process_id=process_id,
+                            )
+        client.save()
+        return Response('上报成功')
 
     @staticmethod
     @api_view(['GET'])
